grpc-demo/grpc_requirements.txt
alembic==1.13.0
annotated-types==0.6.0
ecdsa==0.18.0
greenlet==3.0.1
grpcio==1.59.3
grpcio-tools==1.59.3
importlib-metadata==7.0.0
importlib-resources==6.1.1
Mako==1.3.0
MarkupSafe==2.1.3
passlib==1.7.4
protobuf==4.25.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.5.2
pydantic-core==2.14.5
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.23
typing-extensions==4.8.0
zipp==3.17.0
